Fix: monthCalendar class was deleted, new implementations in sqlite class methods
